+++ tests/src/traffic.rs
@@ -17,8 +17,13 @@ mod traffic_tests {
     use tokio::{fs::File, io::AsyncWriteExt};
 
     use crate::utils::{
-        application::Application, config_dir, hostname_service, ipv6::ipv6_service, kube_client,
-        kube_service::KubeService, run_command::run_exec_with_target, service, udp_logger_service,
+        application::Application,
+        config_dir,
+        ipv6::ipv6_service,
+        kube_client,
+        kube_service::KubeService,
+        run_command::run_exec_with_target,
+        services::{hostname_service, service, udp_logger_service},
         CONTAINER_NAME,
     };
 
@@ -26,7 +31,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(240))]
-    pub async fn remote_dns_enabled_works(#[future] service: KubeService) {
+    pub async fn remote_dns_enabled_works(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec![
             "node",
@@ -49,7 +54,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(240))]
-    pub async fn remote_dns_lookup_google(#[future] service: KubeService) {
+    pub async fn remote_dns_lookup_google(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec![
             "node",
@@ -74,7 +79,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn outgoing_traffic_single_request_enabled(#[future] service: KubeService) {
+    pub async fn outgoing_traffic_single_request_enabled(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec![
             "node",
@@ -97,7 +102,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[should_panic]
-    pub async fn outgoing_traffic_single_request_ipv6(#[future] service: KubeService) {
+    pub async fn outgoing_traffic_single_request_ipv6(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec![
             "node",
@@ -156,7 +161,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn outgoing_traffic_single_request_disabled(#[future] service: KubeService) {
+    pub async fn outgoing_traffic_single_request_disabled(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec![
             "node",
@@ -179,7 +184,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn outgoing_traffic_make_request_after_listen(#[future] service: KubeService) {
+    pub async fn outgoing_traffic_make_request_after_listen(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec![
             "node",
@@ -202,7 +207,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn outgoing_connection_to_self(#[future] service: KubeService) {
+    pub async fn outgoing_connection_to_self(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec!["node", "node-e2e/outgoing/outgoing_connection_to_self.mjs"];
         let mut process = run_exec_with_target(
@@ -226,7 +231,7 @@ mod traffic_tests {
     #[timeout(Duration::from_secs(240))]
     pub async fn outgoing_traffic_udp_with_connect(
         #[future] udp_logger_service: KubeService,
-        #[future] service: KubeService,
+        #[future] services: KubeService,
         #[future] kube_client: Client,
     ) {
         let internal_service = udp_logger_service.await; // Only reachable from withing the cluster.
@@ -311,7 +316,7 @@ mod traffic_tests {
     pub async fn outgoing_traffic_filter_udp_with_connect(
         config_dir: &Path,
         #[future] udp_logger_service: KubeService,
-        #[future] service: KubeService,
+        #[future] services: KubeService,
         #[future] kube_client: Client,
     ) {
         let internal_service = udp_logger_service.await; // Only reachable from withing the cluster.
@@ -425,7 +430,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(30))]
-    pub async fn outgoing_disabled_udp(#[future] service: KubeService) {
+    pub async fn outgoing_disabled_udp(#[future] services: KubeService) {
         let service = service.await;
         // Binding specific port, because if we bind 0 then we get a  port that is bypassed by
         // mirrord and then the tested crash is not prevented by the fix but by the bypassed port.
@@ -468,7 +473,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn go21_outgoing_traffic_single_request_enabled(#[future] service: KubeService) {
+    pub async fn go21_outgoing_traffic_single_request_enabled(#[future] services: KubeService) {
         let command = vec!["go-e2e-outgoing/21.go_test_app"];
         test_go(service, command).await;
     }
@@ -477,7 +482,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(60))]
-    pub async fn go22_outgoing_traffic_single_request_enabled(#[future] service: KubeService) {
+    pub async fn go22_outgoing_traffic_single_request_enabled(#[future] services: KubeService) {
         let command = vec!["go-e2e-outgoing/22.go_test_app"];
         test_go(service, command).await;
     }
@@ -485,7 +490,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn go23_outgoing_traffic_single_request_enabled(#[future] service: KubeService) {
+    pub async fn go23_outgoing_traffic_single_request_enabled(#[future] services: KubeService) {
         let command = vec!["go-e2e-outgoing/23.go_test_app"];
         test_go(service, command).await;
     }
@@ -494,7 +499,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(60))]
-    pub async fn go21_dns_lookup(#[future] service: KubeService) {
+    pub async fn go21_dns_lookup(#[future] services: KubeService) {
         let command = vec!["go-e2e-dns/21.go_test_app"];
         test_go(service, command).await;
     }
@@ -503,7 +508,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(60))]
-    pub async fn go22_dns_lookup(#[future] service: KubeService) {
+    pub async fn go22_dns_lookup(#[future] services: KubeService) {
         let command = vec!["go-e2e-dns/22.go_test_app"];
         test_go(service, command).await;
     }
@@ -512,7 +517,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(60))]
-    pub async fn go23_dns_lookup(#[future] service: KubeService) {
+    pub async fn go23_dns_lookup(#[future] services: KubeService) {
         let command = vec!["go-e2e-dns/23.go_test_app"];
         test_go(service, command).await;
     }
@@ -520,7 +525,7 @@ mod traffic_tests {
     #[cfg_attr(not(feature = "job"), ignore)]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn listen_localhost(#[future] service: KubeService) {
+    pub async fn listen_localhost(#[future] services: KubeService) {
         let service = service.await;
         let node_command = vec!["node", "node-e2e/listen/test_listen_localhost.mjs"];
         let mut process = run_exec_with_target(
@@ -574,7 +579,7 @@ mod traffic_tests {
             "ghcr.io/metalbear-co/mirrord-unix-socket-server:latest",
             "unix-echo"
         )]
-        service: KubeService,
+        services: KubeService,
     ) {
         let service = service.await;
         let app_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
@@ -605,7 +610,7 @@ mod traffic_tests {
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     #[timeout(Duration::from_secs(240))]
-    pub async fn outgoing_bypassed_unix_stream_pathname(#[future] service: KubeService) {
+    pub async fn outgoing_bypassed_unix_stream_pathname(#[future] services: KubeService) {
         let service = service.await;
         let app_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
             .join("../target/debug/rust-bypassed-unix-socket")
@@ -645,7 +650,7 @@ mod traffic_tests {
     #[case::outgoing_disabled(false)]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     pub async fn test_outgoing_traffic_many_requests(
-        #[future] service: KubeService,
+        #[future] services: KubeService,
         #[case] outgoing_enabled: bool,
     ) {
         const HOSTS: &[&str] = &[
